+++ src/test/java/com/mapshot/api/presentation/admin/AdminControllerTest.java
@@ -4,10 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mapshot.api.SlackMockExtension;
 import com.mapshot.api.domain.admin.AdminEntity;
 import com.mapshot.api.domain.admin.AdminRepository;
+import com.mapshot.api.domain.notice.NoticeService;
+import com.mapshot.api.domain.notice.NoticeType;
 import com.mapshot.api.infra.auth.Validation;
-import com.mapshot.api.infra.exception.ApiException;
-import com.mapshot.api.infra.exception.status.ErrorCode;
+import com.mapshot.api.infra.encrypt.EncryptUtil;
 import com.mapshot.api.presentation.admin.model.AdminRequest;
+import com.mapshot.api.presentation.notice.model.NoticeListResponse;
+import com.mapshot.api.presentation.notice.model.NoticeRegistrationRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -20,9 +23,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.test.web.servlet.MockMvc;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
@@ -30,6 +31,8 @@ import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.docu
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -51,6 +54,9 @@ class AdminControllerTest extends SlackMockExtension {
     @Autowired
     private AdminRepository adminRepository;
 
+    @Autowired
+    private NoticeService noticeService;
+
     @Value("${jwt.admin.header}")
     private String ADMIN_HEADER_NAME;
 
@@ -62,34 +68,13 @@ class AdminControllerTest extends SlackMockExtension {
     void init() {
         adminRepository.save(AdminEntity.builder()
                 .nickname("test")
-                .password(encrypt("1234"))
+                .password(EncryptUtil.encrypt("1234"))
                 .build());
     }
 
     @AfterEach
     void release() {
         adminRepository.deleteAll();
-    }
-
-
-    private String encrypt(String text) {
-        try {
-            MessageDigest md = MessageDigest.getInstance(ENCRYPT_ALGORITHM);
-            md.update(text.getBytes());
-
-            return bytesToHex(md.digest());
-        } catch (NoSuchAlgorithmException e) {
-            throw new ApiException(ErrorCode.NO_SUCH_ALGORITHM);
-        }
-
-    }
-
-    private String bytesToHex(byte[] bytes) {
-        StringBuilder builder = new StringBuilder();
-        for (byte b : bytes) {
-            builder.append(String.format("%02x", b));
-        }
-        return builder.toString();
     }
 
 
@@ -115,34 +100,6 @@ class AdminControllerTest extends SlackMockExtension {
                         )));
     }
 
-//    @Test
-//    void 관리자_로그아웃_테스트() throws Exception {
-//
-//
-//        mockMvc.perform(
-//                        RestDocumentationRequestBuilders.post(BASE_URL + "/logout")
-//                                .contentType(MediaType.APPLICATION_JSON)
-//                                .header(ADMIN_HEADER_NAME, adminToken.generate()))
-//                .andExpect(status().isOk())
-//                .andDo(document("admin/logout",
-//                        preprocessRequest(prettyPrint()),
-//                        preprocessResponse(prettyPrint()),
-//                        requestHeaders(
-//                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
-//                        )));
-//    }
-
-//    @Test
-//    void 토큰없이_관리자_로그아웃_요청시_예외() throws Exception {
-//
-//
-//        mockMvc.perform(
-//                        post(BASE_URL + "/logout")
-//                                .contentType(MediaType.APPLICATION_JSON))
-//                .andExpect(status().is4xxClientError());
-//    }
-
-
     @Test
     void 관리자_로그인_연장_테스트() throws Exception {
         mockMvc.perform(
@@ -163,6 +120,126 @@ class AdminControllerTest extends SlackMockExtension {
         mockMvc.perform(
                         post(BASE_URL + "/auth/refresh")
                                 .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().is4xxClientError());
+    }
+
+
+    @Test
+    void 게시글_등록_테스트() throws Exception {
+        NoticeRegistrationRequest request = NoticeRegistrationRequest
+                .builder()
+                .title("헬로")
+                .content("방가방가")
+                .noticeType(NoticeType.FIX.toString())
+                .build();
+
+        mockMvc.perform(
+                        RestDocumentationRequestBuilders.post(BASE_URL + "/notice/register")
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
+                .andExpect(status().isOk())
+                .andDo(document("notice/register",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        requestHeaders(
+                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
+                        ),
+                        requestFields(
+                                fieldWithPath("title").description("공지사항 제목"),
+                                fieldWithPath("content").description("공지사항 내용"),
+                                fieldWithPath("noticeType").description("공지사항 종류")
+                        )));
+    }
+
+    @Test
+    void 관리자가_아닌_사람이_등록_요청시_예외() throws Exception {
+        NoticeRegistrationRequest request = NoticeRegistrationRequest
+                .builder()
+                .title("헬로")
+                .content("방가방가")
+                .noticeType(NoticeType.FIX.toString())
+                .build();
+
+        mockMvc.perform(
+                        post(BASE_URL + "/notice/register")
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().is4xxClientError());
+    }
+
+    @Test
+    void 게시글_삭제_테스트() throws Exception {
+        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
+
+        mockMvc.perform(
+                        RestDocumentationRequestBuilders.get(BASE_URL + "/notice/delete/{noticeNumber}", mostRecentNotice.getId())
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
+                .andExpect(status().isOk())
+                .andDo(document("notice/delete",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        requestHeaders(
+                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("noticeNumber").description("게시글 번호")
+                        )));
+    }
+
+    @Test
+    void 관리자가_아닌_사람이_삭제_요청시_예외() throws Exception {
+        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
+
+        mockMvc.perform(MockMvcRequestBuilders.get(BASE_URL + "/notice/delete/{noticeNumber}", mostRecentNotice.getId())
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().is4xxClientError());
+    }
+
+    @Test
+    void 게시글_수정_테스트() throws Exception {
+        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
+
+        NoticeRegistrationRequest request = NoticeRegistrationRequest.builder()
+                .noticeType(NoticeType.RESERVED_CHECK.toString())
+                .title("수정된 타이틀")
+                .content("수정된 컨텐츠")
+                .build();
+
+        mockMvc.perform(
+                        RestDocumentationRequestBuilders.post(BASE_URL + "/notice/modify/{noticeNumber}", mostRecentNotice.getId())
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
+                .andExpect(status().isOk())
+                .andDo(document("notice/modify",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        requestHeaders(
+                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
+                        ),
+                        requestFields(
+                                fieldWithPath("title").description("수정할 공지사항 제목"),
+                                fieldWithPath("content").description("수정할 공지사항 내용"),
+                                fieldWithPath("noticeType").description("수정할 공지사항 종류")
+                        )));
+    }
+
+
+    @Test
+    void 관리자가_아닌_사람이_수정_요청시_예외() throws Exception {
+        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
+
+        NoticeRegistrationRequest request = NoticeRegistrationRequest.builder()
+                .noticeType(NoticeType.RESERVED_CHECK.toString())
+                .title("수정된 타이틀")
+                .content("수정된 컨텐츠")
+                .build();
+
+        mockMvc.perform(post(BASE_URL + "/notice/modify/{noticeNumber}", mostRecentNotice.getId())
+                        .content(mapper.writeValueAsString(request))
+                        .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().is4xxClientError());
     }
