+++ src/test/java/com/mapshot/api/presentation/notice/NoticeControllerTest.java
@@ -10,7 +10,6 @@ import com.mapshot.api.domain.notice.NoticeType;
 import com.mapshot.api.infra.auth.Validation;
 import com.mapshot.api.presentation.notice.model.NoticeDetailResponse;
 import com.mapshot.api.presentation.notice.model.NoticeListResponse;
-import com.mapshot.api.presentation.notice.model.NoticeRegistrationRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,25 +18,20 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import java.util.Comparator;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
-import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 
@@ -138,124 +132,5 @@ class NoticeControllerTest extends SlackMockExtension {
                 });
 
         assertThat(actual.getId()).isEqualTo(requestId);
-    }
-
-    @Test
-    void 게시글_등록_테스트() throws Exception {
-        NoticeRegistrationRequest request = NoticeRegistrationRequest
-                .builder()
-                .title("헬로")
-                .content("방가방가")
-                .noticeType(NoticeType.FIX.toString())
-                .build();
-
-        mockMvc.perform(
-                        RestDocumentationRequestBuilders.post(BASE_URL + "/register")
-                                .content(mapper.writeValueAsString(request))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
-                .andExpect(status().isOk())
-                .andDo(document("notice/register",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestHeaders(
-                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
-                        ),
-                        requestFields(
-                                fieldWithPath("title").description("공지사항 제목"),
-                                fieldWithPath("content").description("공지사항 내용"),
-                                fieldWithPath("noticeType").description("공지사항 종류")
-                        )));
-    }
-
-    @Test
-    void 관리자가_아닌_사람이_등록_요청시_예외() throws Exception {
-        NoticeRegistrationRequest request = NoticeRegistrationRequest
-                .builder()
-                .title("헬로")
-                .content("방가방가")
-                .noticeType(NoticeType.FIX.toString())
-                .build();
-
-        mockMvc.perform(
-                        post(BASE_URL + "/register")
-                                .content(mapper.writeValueAsString(request))
-                                .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().is4xxClientError());
-    }
-
-    @Test
-    void 게시글_삭제_테스트() throws Exception {
-        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
-
-        mockMvc.perform(
-                        RestDocumentationRequestBuilders.get(BASE_URL + "/delete/{noticeNumber}", mostRecentNotice.getId())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
-                .andExpect(status().isOk())
-                .andDo(document("notice/delete",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestHeaders(
-                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
-                        ),
-                        pathParameters(
-                                parameterWithName("noticeNumber").description("게시글 번호")
-                        )));
-    }
-
-    @Test
-    void 관리자가_아닌_사람이_삭제_요청시_예외() throws Exception {
-        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
-
-        mockMvc.perform(MockMvcRequestBuilders.get(BASE_URL + "/delete/{noticeNumber}", mostRecentNotice.getId())
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().is4xxClientError());
-    }
-
-    @Test
-    void 게시글_수정_테스트() throws Exception {
-        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
-
-        NoticeRegistrationRequest request = NoticeRegistrationRequest.builder()
-                .noticeType(NoticeType.RESERVED_CHECK.toString())
-                .title("수정된 타이틀")
-                .content("수정된 컨텐츠")
-                .build();
-
-        mockMvc.perform(
-                        RestDocumentationRequestBuilders.post(BASE_URL + "/modify/{noticeNumber}", mostRecentNotice.getId())
-                                .content(mapper.writeValueAsString(request))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .headers(HttpHeaders.readOnlyHttpHeaders(adminValidation.makeHeader())))
-                .andExpect(status().isOk())
-                .andDo(document("notice/modify",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestHeaders(
-                                headerWithName(ADMIN_HEADER_NAME).description("관리자 인증 토큰")
-                        ),
-                        requestFields(
-                                fieldWithPath("title").description("수정할 공지사항 제목"),
-                                fieldWithPath("content").description("수정할 공지사항 내용"),
-                                fieldWithPath("noticeType").description("수정할 공지사항 종류")
-                        )));
-    }
-
-
-    @Test
-    void 관리자가_아닌_사람이_수정_요청시_예외() throws Exception {
-        NoticeListResponse mostRecentNotice = noticeService.getNoticeList(0).get(0);
-
-        NoticeRegistrationRequest request = NoticeRegistrationRequest.builder()
-                .noticeType(NoticeType.RESERVED_CHECK.toString())
-                .title("수정된 타이틀")
-                .content("수정된 컨텐츠")
-                .build();
-
-        mockMvc.perform(post(BASE_URL + "/modify/{noticeNumber}", mostRecentNotice.getId())
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().is4xxClientError());
     }
 }
